Major update version 1.2
This version changes a lot! There is now scripting, the binner works properly and different projections can be generated. Masks have been introduced as a way to select parts of reciprocal space to include in the binning. There is also an early draft of a user manual. Several other bugs are also fixed.